provisioner/requirements.txt
Flask==2.3.*
requests==2.31.*
PyJWT[crypto]==2.7.*
google-cloud-firestore==2.11.*
